% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{getClient}
\alias{getClient}
\title{Get OpenReview Client}
\usage{
getClient(username, password)
}
\arguments{
\item{username}{A character vector of length 1.}

\item{password}{A character vector of length 1.}
}
\value{
An instance of the openreview client. Class: 'openreview.openreview.Client',
'python.builtin.object'
}
\description{
\code{getClient} starts a client for openreview. The code is based on:
\url{https://github.com/kevinrue/openreview-r/blob/main/dev/test.R} and
following the instructions in the documentation of the \code{basilisk}
package.
}
\details{
Since user credentials are confidential, it's recommended to get the values
from environment variables rather than directly typing in the console or
script. \link{Sys.getenv} can be useful to get the credential values from
environment variables.
}
\examples{

\dontrun{
username = Sys.getenv("OPENREVIEW_USERNAME")
password = Sys.getenv("OPENREVIEW_PASSWORD")
client = getClient(username = username, password = password)
invitations = getInvitationId(client = client, regex = "Microbiome-VIF.org/.*")
invitations
}

}
