% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getInvitationId}
\alias{getInvitationId}
\title{Get Invitation ID}
\usage{
getInvitationId(client, regex)
}
\arguments{
\item{client}{Client object}

\item{regex}{Character vector of length 1.}
}
\value{
A character vector of all invitations.
}
\description{
\code{getInvitationId} gets invitation ids for a determined regex.
Currently, only the first 1000 items.
}
